light refactoring of internal classes
