+++ OpenAI-DotNet/Audio/AudioEndpoint.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using OpenAI.Extensions;
+using System.IO;
 using System.Net.Http;
 using System.Text.Json;
 using System.Text.Json.Serialization;
