+++ OpenAI-DotNet/Chat/ChatEndpoint.cs
@@ -1,3 +1,4 @@
+using OpenAI.Extensions;
 using System;
 using System.Collections.Generic;
 using System.IO;
