+++ OpenAI-DotNet/Extensions/ResponseExtensions.cs
@@ -7,7 +7,7 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace OpenAI
+namespace OpenAI.Extensions
 {
     internal static class ResponseExtensions
     {
