+++ OpenAI-DotNet/Extensions/StringExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net.Http;
 using System.Text;
 
-namespace OpenAI
+namespace OpenAI.Extensions
 {
     internal static class StringExtensions
     {
