+++ OpenAI-DotNet/Files/FilesEndpoint.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using OpenAI.Extensions;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Net.Http;
